GOT NOTHING but with right calculations

+++ includes/struct.h
@@ -6,7 +6,7 @@
 /*   By: mwilk <mwilk@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/10/07 15:51:38 by mwilk             #+#    #+#             */
-/*   Updated: 2016/10/25 18:27:38 by mwilk            ###   ########.fr       */
+/*   Updated: 2016/10/27 17:12:35 by mwilk            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,9 +26,9 @@ typedef struct			s_color
 
 typedef struct			s_vec3
 {
-	float				x;
-	float				y;
-	float				z;
+	double				x;
+	double				y;
+	double				z;
 }						t_vec3;
 
 struct					s_ray
@@ -42,10 +42,10 @@ typedef struct			s_cam
 	t_vec3				p;
 	t_vec3				vd;
 	t_vec3				upleft;
-	float				w;
-	float				h;
-	float				f;
-	float				fov;
+	double				w;
+	double				h;
+	double				f;
+	double				fov;
 
 }						t_cam;
 
@@ -55,7 +55,7 @@ typedef struct			s_light
 	t_vec3				spotlight;
 	int					type;
 	int					radius;
-	float				angle;
+	double				angle;
 	t_color				color;
 	struct s_light		*next;
 	struct s_light		*prev;
@@ -64,7 +64,7 @@ typedef struct			s_light
 typedef struct			s_sphere
 {
 	t_vec3				p;
-	float				r;
+	double				r;
 	t_color				color;
 }						t_sphere;
 
@@ -78,9 +78,9 @@ typedef struct			s_plane
 
 typedef struct			s_material
 {
-	t_vec3		color;
-	float		diffuse;
-	float		specular;
+	t_vec3				color;
+	double				diffuse;
+	double				specular;
 }						t_material;
 
 
@@ -98,10 +98,10 @@ struct					s_object
 	t_color				ambient;
 	t_color				self_illu;
 	
-	float				shininess;
-	float				shinestrength;
-	float				transmittivity;
-	float				reflectivity;
+	double				shininess;
+	double				shinestrength;
+	double				transmittivity;
+	double				reflectivity;
 	
 	int					permanent;
 	struct s_object		*next;
@@ -110,7 +110,7 @@ struct					s_object
 
 struct					s_hitpoint
 {
-	float				t;
+	double				t;
 	t_vec3				p;
 	t_vec3				n;
 	t_color				c;
